requirements.txt
streamlit
transformers
torch
accelerate
scipy
safetensors
diffusers
Pillow
sqlalchemy